+++ brtp/misc/utils/__init__.py
@@ -1,1 +1,0 @@
-from ._zip_random import zip_random