+++ tests/collections/test_zip_random.py
@@ -1,6 +1,6 @@
 import pytest
 
-from brtp.misc.utils import zip_random
+from brtp.collections import zip_random
 
 
 @pytest.mark.parametrize("seed", [None, 0, 42, 1234])
@@ -13,10 +13,10 @@ def test_zip_random_invariants(seed: int | None):
     result = list(zip_random(i1, i2, seed=seed))
 
     # --- assert ------------------------------------------
-    assert {v for v, _ in result} == set(i1)
-    assert [v for v, _ in result] != i1
-    assert {v for _, v in result} == set(i2)
-    assert [v for _, v in result] != i2
+    assert {v for v, _ in result} == set(i1), "all elements should be present"
+    assert [v for v, _ in result] != i1, "iterable should be randomized"
+    assert {v for _, v in result} == set(i2), "all elements should be present"
+    assert [v for _, v in result] != i2, "iterable should be randomized"
 
 
 def test_zip_random_seed():
@@ -34,4 +34,4 @@ def test_zip_random_seed():
     assert result_1 != result_2
     assert result_1 != result_3
     assert result_2 != result_3
-    assert result_3 == result_4
+    assert result_3 == result_4, "same seed should give same result"